refactor(server): goal 走行を別 TU (mapoi_nav2_bridge_goal.cpp) へ分離
#345 第1弾。NavigateToPose (単発 Go) のメソッド定義を別ファイルへ移す
のみで、クラス構造・挙動は変更しない。
